+++ src/ymf262/topcymbaloperator.h
@@ -15,14 +15,14 @@ public:
 	TopCymbalOperator( Opl3* opl, int baseAddress = topCymbalOperatorBaseAddress ) : Operator( opl, baseAddress ) {
 	}
 
-	double getOperatorOutput( double modulator ) ;
+	double getOperatorOutput( int modulator ) ;
 
 protected:
 	// This method is used here with the HighHatOperator phase
 	// as the externalPhase.
 	// Conversely, this method is also used through inheritance by the HighHatOperator,
 	// now with the TopCymbalOperator phase as the externalPhase.
-	double getOperatorOutput( double modulator, double externalPhase ) ;
+	double getOperatorOutput( int modulator, int externalPhase ) ;
 };
 }
